+++ shared/include/shared/world.h
@@ -14,6 +14,6 @@ struct world
 void world_load(struct world *world, const char *filename, bool load_maps);
 void world_unload(struct world *world, bool unload_maps);
 
-int world_get_map_index(struct world *world, const char *filename);
+size_t world_get_map_index(struct world *world, const char *filename);
 
 #endif
